+++ pyquil/api/_qpu_client.py
@@ -16,6 +16,7 @@
 from dataclasses import dataclass
 from datetime import datetime
 from typing import Dict, Optional, cast, Tuple, Union, List, Any
+from attr import field
 
 import rpcq
 from dateutil.parser import parse as parsedate
@@ -93,6 +94,9 @@ class GetBuffersResponse:
     buffers: Dict[str, BufferResponse]
     """Job buffers, by buffer name."""
 
+    execution_duration_microseconds: Optional[int] = field(default=None)
+    "Duration job held exclusive hardware access."
+
 
 class QPUClient:
     """
@@ -154,6 +158,27 @@ class QPUClient:
                 )
                 for name, val in buffs.items()
             }
+        )
+
+    def get_execution_results(self, request: GetBuffersRequest) -> GetBuffersResponse:
+        """
+        Get job buffers and execution metadata.
+        """
+        result = self._rpcq_request(
+            "get_execution_results",
+            job_id=request.job_id,
+            wait=request.wait,
+        )
+        return GetBuffersResponse(
+            buffers={
+                name: BufferResponse(
+                    shape=cast(Tuple[int, int], tuple(val["shape"])),
+                    dtype=val["dtype"],
+                    data=val["data"],
+                )
+                for name, val in result.buffers.items()
+            },
+            execution_duration_microseconds=result.execution_duration_microseconds,
         )
 
     @retry(exceptions=TimeoutError, tries=2)  # type: ignore